Fix blog page compilation - separate server and client components for proper SSG
