+++ components/BlogPageClient.tsx
@@ -1,0 +1,260 @@
+"use client";
+
+import { useState, useMemo } from 'react';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { Badge } from '@/components/ui/badge';
+import { Input } from '@/components/ui/input';
+import { Button } from '@/components/ui/button';
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from '@/components/ui/select';
+import { Calendar, Clock, User, ArrowRight, FileText, Search, X } from 'lucide-react';
+import { format } from 'date-fns';
+import Link from 'next/link';
+import Image from 'next/image';
+
+interface BlogPost {
+  slug: string;
+  title: string;
+  description?: string;
+  date: string;
+  author: string;
+  category?: string;
+  tags?: string[];
+  image?: string;
+  readingTime: string;
+}
+
+interface BlogPageClientProps {
+  posts: BlogPost[];
+}
+
+export function BlogPageClient({ posts }: BlogPageClientProps) {
+  const [searchQuery, setSearchQuery] = useState("");
+  const [selectedCategory, setSelectedCategory] = useState<string>("all");
+
+  // Extract unique categories from posts
+  const categories = useMemo(() => {
+    const cats = new Set<string>();
+    posts.forEach(post => {
+      if (post.category) {
+        cats.add(post.category);
+      }
+    });
+    return Array.from(cats).sort();
+  }, [posts]);
+
+  // Filter posts based on search and category
+  const filteredPosts = useMemo(() => {
+    return posts.filter(post => {
+      const matchesSearch = searchQuery === "" || 
+        post.title.toLowerCase().includes(searchQuery.toLowerCase()) ||
+        (post.description && post.description.toLowerCase().includes(searchQuery.toLowerCase()));
+      
+      const matchesCategory = selectedCategory === "all" || 
+        post.category === selectedCategory;
+      
+      return matchesSearch && matchesCategory;
+    });
+  }, [posts, searchQuery, selectedCategory]);
+
+  const clearFilters = () => {
+    setSearchQuery("");
+    setSelectedCategory("all");
+  };
+
+  const hasActiveFilters = searchQuery !== "" || selectedCategory !== "all";
+
+  return (
+    <section className="container py-12 md:py-20">
+      <div className="mx-auto max-w-6xl">
+        {/* Search and Filter Controls */}
+        <div className="mb-8">
+          <div className="flex flex-col sm:flex-row gap-4">
+            {/* Search Input */}
+            <div className="relative flex-1">
+              <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 h-4 w-4 text-muted-foreground" />
+              <Input
+                type="text"
+                placeholder="Search posts..."
+                value={searchQuery}
+                onChange={(e) => setSearchQuery(e.target.value)}
+                className="pl-10"
+              />
+            </div>
+
+            {/* Category Dropdown */}
+            <Select value={selectedCategory} onValueChange={setSelectedCategory}>
+              <SelectTrigger className="w-full sm:w-[200px]">
+                <SelectValue placeholder="All Categories" />
+              </SelectTrigger>
+              <SelectContent>
+                <SelectItem value="all">All Categories</SelectItem>
+                {categories.map(category => (
+                  <SelectItem key={category} value={category}>
+                    {category}
+                  </SelectItem>
+                ))}
+              </SelectContent>
+            </Select>
+
+            {/* Clear Filters Button */}
+            {hasActiveFilters && (
+              <Button
+                variant="outline"
+                size="default"
+                onClick={clearFilters}
+                className="flex items-center gap-2"
+              >
+                <X className="h-4 w-4" />
+                Clear
+              </Button>
+            )}
+          </div>
+
+          {/* Active Filters Display */}
+          {hasActiveFilters && (
+            <div className="mt-4 flex flex-wrap items-center gap-2">
+              <span className="text-sm text-muted-foreground">Active filters:</span>
+              {searchQuery && (
+                <Badge variant="secondary" className="flex items-center gap-1">
+                  Search: &quot;{searchQuery}&quot;
+                  <button
+                    onClick={() => setSearchQuery("")}
+                    className="ml-1 hover:text-destructive"
+                  >
+                    <X className="h-3 w-3" />
+                  </button>
+                </Badge>
+              )}
+              {selectedCategory !== "all" && (
+                <Badge variant="secondary" className="flex items-center gap-1">
+                  Category: {selectedCategory}
+                  <button
+                    onClick={() => setSelectedCategory("all")}
+                    className="ml-1 hover:text-destructive"
+                  >
+                    <X className="h-3 w-3" />
+                  </button>
+                </Badge>
+              )}
+            </div>
+          )}
+
+          {/* Results Count */}
+          <div className="mt-4 text-sm text-muted-foreground">
+            Showing {filteredPosts.length} of {posts.length} posts
+          </div>
+        </div>
+
+        {filteredPosts.length === 0 ? (
+          <div className="py-20 text-center">
+            <p className="text-lg text-muted-foreground">
+              No posts found matching your criteria.
+            </p>
+            {hasActiveFilters && (
+              <Button
+                variant="outline"
+                onClick={clearFilters}
+                className="mt-4"
+              >
+                Clear filters
+              </Button>
+            )}
+          </div>
+        ) : (
+          <div className="grid gap-8 md:grid-cols-2 lg:grid-cols-3">
+            {filteredPosts.map((post) => {
+              // Use the slug from frontmatter (now required)
+              let postUrl = post.slug;
+              if (!postUrl.startsWith('/')) {
+                postUrl = `/${postUrl}`;
+              }
+              if (!postUrl.endsWith('/')) {
+                postUrl += '/';
+              }
+              
+              return (
+                <Link key={post.slug} href={postUrl}>
+                  <Card className="h-full transition-all hover:shadow-lg hover:-translate-y-1 overflow-hidden">
+                    {post.image ? (
+                      <div className="aspect-video w-full overflow-hidden relative">
+                        <Image
+                          src={post.image}
+                          alt={post.title}
+                          fill
+                          className="object-cover"
+                          sizes="(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw"
+                        />
+                      </div>
+                    ) : (
+                      <div className="aspect-video w-full bg-gradient-to-br from-primary/10 to-primary/5 flex items-center justify-center">
+                        <div className="text-center p-6">
+                          <div className="inline-flex h-16 w-16 items-center justify-center rounded-full bg-primary/10 mb-4">
+                            <FileText className="h-8 w-8 text-primary/60" />
+                          </div>
+                          <div className="flex flex-wrap gap-2 justify-center">
+                            {post.tags && Array.isArray(post.tags) && post.tags.slice(0, 2).map((tag: string) => (
+                              <Badge key={tag} variant="secondary" className="text-xs">
+                                {tag}
+                              </Badge>
+                            ))}
+                          </div>
+                        </div>
+                      </div>
+                    )}
+                    <CardHeader>
+                      {post.category && (
+                        <button
+                          onClick={(e) => {
+                            e.preventDefault();
+                            setSelectedCategory(post.category!);
+                          }}
+                          className="inline-block"
+                        >
+                          <Badge className="mb-2 cursor-pointer hover:bg-secondary/80 transition-colors" variant="secondary">
+                            {post.category}
+                          </Badge>
+                        </button>
+                      )}
+                      <CardTitle className="line-clamp-2">{post.title}</CardTitle>
+                      <CardDescription className="line-clamp-3">
+                        {post.description}
+                      </CardDescription>
+                    </CardHeader>
+                    <CardContent>
+                      <div className="flex flex-wrap items-center gap-4 text-xs text-muted-foreground">
+                        <div className="flex items-center gap-1">
+                          <User className="h-3 w-3" />
+                          <span>{post.author}</span>
+                        </div>
+                        <div className="flex items-center gap-1">
+                          <Calendar className="h-3 w-3" />
+                          <time dateTime={post.date}>
+                            {format(new Date(post.date), 'MMM d, yyyy')}
+                          </time>
+                        </div>
+                        <div className="flex items-center gap-1">
+                          <Clock className="h-3 w-3" />
+                          <span>{post.readingTime}</span>
+                        </div>
+                      </div>
+                      <div className="mt-4 flex items-center text-sm font-medium text-primary">
+                        Read more
+                        <ArrowRight className="ml-1 h-4 w-4" />
+                      </div>
+                    </CardContent>
+                  </Card>
+                </Link>
+              );
+            })}
+          </div>
+        )}
+      </div>
+    </section>
+  );
+}
